+++ R/utils.R
@@ -12,33 +12,3 @@
 #' @param rhs A function call using the magrittr semantics.
 #' @return The result of calling `rhs(lhs)`.
 NULL
-
-# Borrowed from DatabaseConnector: https://github.com/OHDSI/DatabaseConnector/blob/6bdece864cfa9a5c9ea9b6d5473df4d2f8a30335/R/DatabaseConnector.R#L75
-
-# Borrowed from devtools: https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L44
-is_installed <- function(pkg, version = 0) {
-  installed_version <- tryCatch(utils::packageVersion(pkg),
-                                error = function(e) NA
-  )
-  !is.na(installed_version) && installed_version >= version
-}
-
-# Borrowed and adapted from devtools: https://github.com/hadley/devtools/blob/ba7a5a4abd8258c52cb156e7b26bb4bf47a79f0b/R/utils.r#L74
-
-#' @importFrom rlang inform
-#' @importFrom utils install.packages menu
-ensure_installed <- function(pkg) {
-  if (!is_installed(pkg)) {
-    msg <- paste0(sQuote(pkg), " must be installed for this functionality.")
-    if (interactive()) {
-      rlang::inform(paste(msg, "Would you like to install it?", sep = "\n"))
-      if (utils::menu(c("Yes", "No")) == 1) {
-        utils::install.packages(pkg)
-      } else {
-        stop(msg, call. = FALSE)
-      }
-    } else {
-      stop(msg, call. = FALSE)
-    }
-  }
-}