+++ NEWS.md
@@ -1,3 +1,12 @@
+# CDMConnector 0.3.0
+
+* Add computePermanent and appendPermanent functions
+* Add readCohortSet
+* Add generateCohortSet
+* Extract Eunomia dataset so that it is downloaded separately
+
+Released Nov 29, 2022
+
 # CDMConnector 0.2.0
 
 * Add support for OMOP v5.3 and v5.4